// DogdexAI/src/lib/types.ts
export type CollectionSource = 'image_upload' | 'video_upload' | 'stream_capture';
// ...
export interface PredictionHistoryItem {
  id: string;
  processedMediaUrl: string;
  modelUsed: string;
  isCorrect?: boolean | null;
  createdAt: string;
  updatedAt: string;
  // Detections đã được tinh gọn cho trang danh sách
  detections: {
    detectedBreed: string;
    breedName: string;
    confidence: number;
  }[];
  media: { name: string; mediaUrl: string; type: string };
  source: CollectionSource;
}